% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AnnotatedBK}
\alias{AnnotatedBK}
\title{Annotated Brothers K}
\format{
\subsection{\code{AnnotatedBK}}{

A list with 3 elements:
\describe{
\item{token}{Contains token information}
\item{entity}{Contains entity information}
\item{document}{Contains document information}
}
}
}
\source{
\url{https://www.gutenberg.org/ebooks/28054}
}
\usage{
AnnotatedBK
}
\description{
An annotated version of "The Brothers Karamazov."  See the cleanNLP package
for more information on what this entails.
}
\keyword{datasets}
